+++ client/src/components/Drawer/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Drawer";
