+++ client/src/pages/Home/Home.js
@@ -2,8 +2,8 @@ import React, { Component } from "react";
 import ScrollableTabsButtonAuto from '../../components/Tabs/Tabs';
 import Jumbotron from "../../components/Jumbotron";
 import Grid from '@material-ui/core/Grid';
-import LabelBottomNavigation from '../../components/Footer/Footer'
-
+import LabelBottomNavigation from '../../components/Footer/Footer';
+import PermanentDrawer from '../../components/Drawer/Drawer'
 //import { Table } from "@material-ui/core";
 
 class Home extends Component {
@@ -36,15 +36,16 @@ class Home extends Component {
   render() {
     return (
 	<div>
-  <Jumbotron />
-   <Grid container>
-   <Grid item xs={12} md={3}>
-   </Grid>
-    <Grid item xs={12} md={9}>
-			    <ScrollableTabsButtonAuto />
-    </Grid>
-  </Grid>
-		      <LabelBottomNavigation />
+  		<Jumbotron />
+   			<Grid container>
+   				<Grid item md={3}>
+				   <PermanentDrawer />
+   				</Grid>
+    			<Grid item xs={12} md={9}>
+			    	<ScrollableTabsButtonAuto />
+    			</Grid>
+  			</Grid>
+		<LabelBottomNavigation />
 	</div>
     );
   }
